Move Character classes to their own package
