+++ src/main/kotlin/profile/character/ActiveClassJob.kt
@@ -1,3 +1,3 @@
-package cloud.drakon.ktlodestone.profile
+package cloud.drakon.ktlodestone.profile.character
 
 data class ActiveClassJob(val name: String, val level: Byte)
