+++ src/main/kotlin/profile/character/GrandCompany.kt
@@ -1,3 +1,3 @@
-package cloud.drakon.ktlodestone.profile
+package cloud.drakon.ktlodestone.profile.character
 
 data class GrandCompany(val name: String, val rank: String, val icon: String)
