+++ src/main/kotlin/profile/character/Guardian.kt
@@ -1,3 +1,3 @@
-package cloud.drakon.ktlodestone.profile
+package cloud.drakon.ktlodestone.profile.character
 
 data class Guardian(val name: String, val icon: String)
