+++ src/main/kotlin/profile/character/Town.kt
@@ -1,3 +1,3 @@
-package cloud.drakon.ktlodestone.profile
+package cloud.drakon.ktlodestone.profile.character
 
 data class Town(val name: String, val icon: String)
